Cleaning things up a little.
