+++ macos/Bookmarks/Views/Sidebar.swift
@@ -24,46 +24,7 @@ import SwiftUI
 import BookmarksCore
 import Interact
 
-extension String {
-
-    var favoriteId: BookmarksSection { .favorite(tag: self) }
-    var tagId: BookmarksSection { .tag(tag: self) }
-
-}
-
-struct SidebarLink: View {
-
-    var selection: Binding<BookmarksSection?>
-    var tag: BookmarksSection
-    var title: String
-    var systemImage: String
-    var databaseView: DatabaseView
-
-    func selectionActiveBinding(_ tag: BookmarksSection) -> Binding<Bool> {
-        return Binding {
-            selection.wrappedValue == tag
-        } set: { value in
-            guard value == true else {
-                return
-            }
-            selection.wrappedValue = tag
-        }
-    }
-
-    var body: some View {
-        NavigationLink(destination: ContentView(sidebarSelection: selection, databaseView: databaseView)
-                        .navigationTitle(title),
-                       isActive: selectionActiveBinding(tag)) {
-            Label(title, systemImage: systemImage)
-        }
-        .tag(tag)
-    }
-
-}
-
 struct Sidebar: View {
-
-    // TOOD: Maybe the sidebar selection is an app manager thing, given it's probably top-level app state?
 
     enum SheetType {
         case rename(tag: String)
@@ -88,88 +49,97 @@ struct Sidebar: View {
     }
 
     var body: some View {
-        List(selection: $selection) {
-            Section {
-
-                SidebarLink(selection: $selection,
-                            tag: .all,
-                            title: "All Bookmarks",
-                            systemImage: "bookmark",
-                            databaseView: DatabaseView(database: manager.database))
-
-                SidebarLink(selection: $selection,
-                            tag: .untagged,
-                            title: "Untagged",
-                            systemImage: "tag",
-                            databaseView: DatabaseView(database: manager.database))
-
-            }
-            Section(header: Text("Favourites")) {
-                ForEach(settings.favoriteTags.sorted(), id: \.favoriteId) { tag in
+        ScrollViewReader { scrollView in
+            List(selection: $selection) {
+                Section {
 
                     SidebarLink(selection: $selection,
-                                tag: tag.favoriteId,
-                                title: tag,
-                                systemImage: "tag",
-                                databaseView: DatabaseView(database: manager.database, tags: [tag]))
-                        .contextMenu(ContextMenu(menuItems: {
-                            Button("Remove from Favourites") {
-                                settings.favoriteTags = settings.favoriteTags.filter { $0 != tag }
-                            }
-                            Divider()
-                            Button("Edit on Pinboard") {
-                                do {
-                                    guard let user = manager.user else {
-                                        return
-                                    }
-                                    NSWorkspace.shared.open(try tag.pinboardUrl(for: user))
-                                } catch {
-                                    print("Failed to open on Pinboard error \(error)")
-                                }
-                            }
-                        }))
-
-                }
-            }
-            Section(header: Text("Tags")) {
-                ForEach(tagsView.tags, id: \.tagId) { tag in
+                                tag: .all,
+                                title: "All Bookmarks",
+                                systemImage: "bookmark",
+                                databaseView: DatabaseView(database: manager.database))
 
                     SidebarLink(selection: $selection,
-                                tag: tag.tagId,
-                                title: tag,
+                                tag: .untagged,
+                                title: "Untagged",
                                 systemImage: "tag",
-                                databaseView: DatabaseView(database: manager.database, tags: [tag]))
-                        .contextMenu(ContextMenu(menuItems: {
-                            Button("Rename") {
-                                self.sheet = .rename(tag: tag)
-                            }
-                            Button("Delete") {
-                                self.manager.pinboard.tags_delete(tag) { _ in
-                                    self.manager.updater.start()
-                                }
-                            }
-                            Divider()
-                            Button("Add to Favourites") {
-                                var favoriteTags = settings.favoriteTags
-                                favoriteTags.append(tag)
-                                settings.favoriteTags = favoriteTags
-                            }
-                            Divider()
-                            Button("Edit on Pinboard") {
-                                do {
-                                    guard let user = manager.user else {
-                                        return
-                                    }
-                                    NSWorkspace.shared.open(try tag.pinboardUrl(for: user))
-                                } catch {
-                                    print("Failed to open on Pinboard error \(error)")
-                                }
-                            }
-                        }))
+                                databaseView: DatabaseView(database: manager.database))
 
                 }
-            }
+                Section(header: Text("Favourites")) {
+                    ForEach(settings.favoriteTags.sorted(), id: \.favoriteId) { tag in
 
+                        SidebarLink(selection: $selection,
+                                    tag: tag.favoriteId,
+                                    title: tag,
+                                    systemImage: "tag",
+                                    databaseView: DatabaseView(database: manager.database, tags: [tag]))
+                            .contextMenu(ContextMenu(menuItems: {
+                                Button("Remove from Favourites") {
+                                    settings.favoriteTags = settings.favoriteTags.filter { $0 != tag }
+                                }
+                                Divider()
+                                Button("Edit on Pinboard") {
+                                    do {
+                                        guard let user = manager.user else {
+                                            return
+                                        }
+                                        NSWorkspace.shared.open(try tag.pinboardUrl(for: user))
+                                    } catch {
+                                        print("Failed to open on Pinboard error \(error)")
+                                    }
+                                }
+                            }))
+
+                    }
+                }
+                Section(header: Text("Tags")) {
+                    ForEach(tagsView.tags, id: \.tagId) { tag in
+
+                        SidebarLink(selection: $selection,
+                                    tag: tag.tagId,
+                                    title: tag,
+                                    systemImage: "tag",
+                                    databaseView: DatabaseView(database: manager.database, tags: [tag]))
+                            .contextMenu(ContextMenu(menuItems: {
+                                Button("Rename") {
+                                    self.sheet = .rename(tag: tag)
+                                }
+                                Button("Delete") {
+                                    self.manager.pinboard.tags_delete(tag) { _ in
+                                        self.manager.updater.start()
+                                    }
+                                }
+                                Divider()
+                                Button("Add to Favourites") {
+                                    var favoriteTags = settings.favoriteTags
+                                    favoriteTags.append(tag)
+                                    settings.favoriteTags = favoriteTags
+                                }
+                                Divider()
+                                Button("Edit on Pinboard") {
+                                    do {
+                                        guard let user = manager.user else {
+                                            return
+                                        }
+                                        NSWorkspace.shared.open(try tag.pinboardUrl(for: user))
+                                    } catch {
+                                        print("Failed to open on Pinboard error \(error)")
+                                    }
+                                }
+                            }))
+
+                    }
+                }
+
+            }
+            .onChange(of: selection) { selection in
+                guard let selection = selection else {
+                    return
+                }
+                print("scrolling to \(selection)")
+                scrollView.scrollTo(selection)
+            }
         }
         .sheet(item: $sheet) { sheet in
             switch sheet {
